remove line item's todo, move this class to a common place
