+++ lib/com/fluidpay/invoice/response.dart
@@ -1,4 +1,5 @@
 import 'package:fluidpay/com/fluidpay/common/base.dart';
+import 'package:fluidpay/com/fluidpay/line_item/line_item.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'common.dart';
