+++ test/com/fluidpay/invoice/request_test.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import 'package:fluidpay/com/fluidpay/invoice/common.dart';
 import 'package:fluidpay/com/fluidpay/invoice/request.dart';
+import 'package:fluidpay/com/fluidpay/line_item/line_item.dart';
 import 'package:test/test.dart';
 
 void main() {
